CRUD estacionamento

+++ src/main/java/com/senai/SA/repository/UsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
 
     Optional<UsuarioModel> findByEmail(String email);
